separate file for handling learnings, separate functions for calling api and saving to db
