+++ apps/lv-pyapi/main.py
@@ -1,4 +1,4 @@
-from fastapi import FastAPI, Depends, HTTPException, Body, BackgroundTasks
+from fastapi import FastAPI, Depends, HTTPException, Body
 from sqlalchemy.orm import Session
 from sqlalchemy import select
 from typing import List
@@ -73,44 +73,6 @@ def get_gemini_response(prompt: str = Body(..., embed=True)):
     except Exception as e:
         return JSONResponse(status_code=500, content={"message": str(e), "status": 500})
 
-
-def learnings_from_messages(messages: List[str]):
-    """Generate learnings from messages using Gemini API"""
-    try:
-        prompt = (
-            "Extract key learnings from the following messages:\n\n" +
-            "\n".join(f"- {msg}" for msg in messages) +
-            "\n\nProvide key insights of the main points."
-        )
-        response = client.models.generate_content(
-            model="gemini-2.5-flash",
-            contents=prompt
-        )
-        return response.text # Should be stored in the database
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Gemini API error: {str(e)}")
-
-
-# Should be triggered when every fifth message is added
-def get_messages_for_learnings(user_id: str, background: BackgroundTasks,  db: Session = Depends(get_db)):
-    """Fetch messages for new learnings"""
-    try:
-        # Query the last 5 messages from the user
-        stmt = (
-            select(ConversationMessage)
-            .where(ConversationMessage.userId == user_id)
-            .order_by(ConversationMessage.createdAt.desc())
-            .limit(5)
-        )
-        result = db.execute(stmt)
-        messages = result.scalars().all()
-        contents = [message.content for message in messages]
-
-        background.add_task(learnings_from_messages, contents) # Process in background
-
-        return {"status": "Learning task initiated"}
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=f"Error creating learnings: {str(e)}")
 
 if __name__ == "__main__":
     uvicorn.run(app, host="0.0.0.0", port=8080)
